audio test

+++ main.cpp
@@ -3,8 +3,8 @@
 // need argc and argv for windows compatibility
 int main(int argc, char *argv[])
 {
-    constexpr size_t WIDTH = 1920 * 0.5;
-    constexpr size_t HEIGHT = 1080 * 0.5;
+    constexpr int WIDTH = 1920 * 0.5;
+    constexpr int HEIGHT = 1080 * 0.5;
 
     APPLICATION app(WIDTH, HEIGHT);
     app.run();
